old-tests/webapps/XMLHttpRequest/tests/submissions/Microsoft/response-data-arraybuffer.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="Help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-response-attribute" />
    <script src="/resources/testharness.js"></script>
    <title>XMLHttpRequest L2: The response attribute: ArrayBuffer data</title>
</head>

<body>
    <div id="log"></div>

    <script type="text/javascript">
        var test = async_test();

        test.step(function()
        {
            var xhr = new XMLHttpRequest();

            xhr.onreadystatechange = function() 
            {
                if (xhr.readyState == 4) 
                {
                    test.step(function() 
                    {
                        assert_equals(xhr.status, 200);

                        var buf = xhr.response;
                        assert_true(buf instanceof ArrayBuffer);

                        var arr = new Uint8Array(buf);
                        assert_equals(arr.length, 5);
                        assert_equals(arr[0], 0x48, "Expect 'H'");
                        assert_equals(arr[1], 0x65, "Expect 'e'");
                        assert_equals(arr[2], 0x6c, "Expect 'l'");
                        assert_equals(arr[3], 0x6c, "Expect 'l'");
                        assert_equals(arr[4], 0x6f, "Expect 'o'");

                        test.done();
                    });
                }
            };

            xhr.open("GET", "./support/echo.php?content=Hello", true);
            xhr.responseType = "arraybuffer";
            xhr.send();
        });
    </script>
</body>
</html>
